requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

sqlalchemy==2.0.25
asyncpg==0.29.0

pydantic==2.5.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

httpx==0.26.0
websockets==12.0

rich==13.7.0
textual==0.52.1
